pyproject.toml
[tool.poetry]
name = "serverpinger"
version = "0.1.0"
description = ""
authors = ["Marcelo Trylesinski <marcelotryle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.61.2"
fastapi-utils = "^0.2.1"
slackclient = "^2.9.3"
uvicorn = "^0.12.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
